% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_opart_gaussian.R
\name{labelled_opart_gaussian}
\alias{labelled_opart_gaussian}
\title{compute the optimal changepoint model for a vector of real-valued data
and a non-negative real-valued penalty,
given the square loss (to minimize) / gaussian likelihood (to maximize)}
\usage{
labelled_opart_gaussian(data, labels, penalty)
}
\arguments{
\item{data}{A numerical vector for which the changepoint model is to be computed}

\item{labels}{A data frame containing information of labelled regions and breaks}

\item{penalty}{A non-negative real number indicating penalty parameter}
}
\value{
A vector of the optimal cost values and a vector of the optimal segment ends
}
\description{
compute the optimal changepoint model for a vector of real-valued data
and a non-negative real-valued penalty,
given the square loss (to minimize) / gaussian likelihood (to maximize)
}
